Add JavaDoc for all class and function Update CodeStandard
