--- src/main/java/Xan/Event.java
@@ -1,17 +1,0 @@
-package Xan;
-
-public class Event extends Task {
-    protected String startTime;
-    protected String endTime;
-
-    public Event(String description, String startTime, String endTime) {
-        super(description);
-        this.startTime = startTime;
-        this.endTime = endTime;
-    }
-
-    @Override
-    public String toString() {
-        return "[E]" + super.toString() + " (from: " + startTime + " to: " + endTime + ")";
-    }
-}
