--- src/main/java/Xan/Todo.java
@@ -1,13 +1,0 @@
-package Xan;
-
-public class Todo extends Task {
-
-    public Todo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
