+++ src/main/java/xan/exception/XanException.java
@@ -1,4 +1,4 @@
-package Xan.exception;
+package xan.exception;
 
 public class XanException extends IllegalArgumentException {
     public XanException(String message) {
