+++ src/main/java/xan/ui/Ui.java
@@ -1,4 +1,4 @@
-package Xan.ui;
+package xan.ui;
 
 public class Ui {
