+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-Hello! I'm Xan! I am your personalized Task manager!
+Hello! I'm xan! I am your personalized Task manager!
 What can I do for you?
 Got it. I've added this task:
 [T][ ] borrow book
